+++ dec/CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this crate will be documented in this file.
 The format is based on [Keep a Changelog], and this crate adheres to [Semantic
 Versioning].
 
+## Unreleased
+
+* Remove the `arbitrary-precision` feature. The arbitrary-precision `Decimal`
+  type is now always compiled.
+
 ## 0.3.3 - 2021-03-10
 
 * Fix bug that could cause `to_standard_notation_string` to misplace the decimal